fix: update cachix action version and ensure interface index is cast to u32

+++ src/iface.rs
@@ -535,7 +535,7 @@ fn do_recvmsg_v4(fd: RawFd) -> io::Result<Datagram> {
     if let Ok(iter) = msg.cmsgs() {
         for c in iter {
             if let ControlMessageOwned::Ipv4PacketInfo(p) = c {
-                recv_ifindex = Some(p.ipi_ifindex);
+                recv_ifindex = Some(p.ipi_ifindex as u32);
             }
         }
     }
@@ -572,7 +572,7 @@ fn do_recvmsg_v6(fd: RawFd) -> io::Result<Datagram> {
     if let Ok(iter) = msg.cmsgs() {
         for c in iter {
             if let ControlMessageOwned::Ipv6PacketInfo(p) = c {
-                recv_ifindex = Some(p.ipi6_ifindex);
+                recv_ifindex = Some(p.ipi6_ifindex as u32);
             }
         }
     }
